Blue Bank Analysis
This is my Python script for Blue Bank Analysis
Blue Bank is a bank in USA that has a loan department which is currently understaffed. They supply loans to individuals and don’t have much reporting on how risky these borrowers are. Using Python and Tableau, they’d like to see a report of borrowers who may have issues paying
back the loan.